fix Homemade Apple font optimization
