+++ src/fonts.js
@@ -1,0 +1,13 @@
+import { Koulen, Homemade_Apple as HomemadeApple } from "next/font/google";
+
+export const koulen = Koulen({
+  weight: "400",
+  subsets: ["latin"],
+  variable: "--font-koulen",
+});
+
+export const homemadeApple = HomemadeApple({
+  weight: "400",
+  subsets: ["latin"],
+  variable: "--font-homemade-apple",
+});
